+++ CHANGES.rst
@@ -3,14 +3,26 @@ Version history
 
 **UNRELEASED**
 
-- Changed nullable relationships to include ``Optional`` in their type annotations
+- Type annotations for ARRAY column attributes now include the Python type of
+  the array elements
+- Added support for specifying engine arguments via ``--engine-arg``
+  (PR by @LajosCseppento)
+- Fixed incorrect package name used in ``importlib.metadata.version`` for
+  ``sqlalchemy-citext``, resolving ``PackageNotFoundError`` (PR by @oaimtiaz)
+- Prevent double pluralization (PR by @dkratzert)
+- Fixes DOMAIN extending JSON/JSONB data types (PR by @sheinbergon)
+
+**3.0.0**
+
 - Dropped support for Python 3.8
+- Changed nullable relationships to include ``Optional`` in their type annotations
 - Fixed SQLModel code generation
 - Fixed two rendering issues in ``ENUM`` columns when a non-default schema is used: an
   unwarranted positional argument and missing the ``schema`` argument
 - Fixed ``AttributeError`` when metadata contains user defined column types
 - Fixed ``AssertionError`` when metadata contains a column type that is a type decorator
   with an all-uppercase name
+- Fixed MySQL ``DOUBLE`` column types being rendered with the wrong arguments
 
 **3.0.0rc5**